// client/src/lib/graphql/mutations/DisConnectStripe/__generated__/DisConnectStripe.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: DisConnectStripe
// ====================================================

export interface DisConnectStripe_disConnectStripe {
  __typename: "Viewer";
  hasWallet: boolean | null;
}

export interface DisConnectStripe {
  disConnectStripe: DisConnectStripe_disConnectStripe;
}